Fetch Unique index information from target db and use that - For Oracle Source db type using partition-by-table strategy completely
