+++ yb-voyager/src/srcdb/oracle.go
@@ -19,11 +19,9 @@ import (
 	"context"
 	"database/sql"
 	"fmt"
-	"sort"
 	"strings"
 	"time"
 
-	goerrors "github.com/go-errors/errors"
 	"github.com/google/uuid"
 	"github.com/samber/lo"
 	log "github.com/sirupsen/logrus"
@@ -628,123 +626,6 @@ func (ora *Oracle) GetServers() []string {
 
 func (ora *Oracle) GetPartitions(tableName sqlname.NameTuple) []string {
 	panic("not implemented")
-}
-
-const oraQueryTmplForUniqIndexes = `
-WITH unique_constraints AS (
-    SELECT
-        consCols.TABLE_NAME,
-        consCols.CONSTRAINT_NAME AS index_key,
-        consCols.COLUMN_NAME,
-        consCols.POSITION AS ordinal_position
-    FROM
-        ALL_CONS_COLUMNS consCols
-    JOIN
-        ALL_CONSTRAINTS cons ON cons.CONSTRAINT_NAME = consCols.CONSTRAINT_NAME
-    WHERE
-        cons.CONSTRAINT_TYPE = 'U'
-        AND cons.OWNER = '%s'
-        AND consCols.TABLE_NAME IN ('%s')
-),
-unique_indexes AS (
-    SELECT
-        indCols.TABLE_NAME,
-        indCols.INDEX_NAME AS index_key,
-        indCols.COLUMN_NAME,
-        indCols.COLUMN_POSITION AS ordinal_position
-    FROM
-        ALL_IND_COLUMNS indCols
-    JOIN
-        ALL_INDEXES ind ON ind.INDEX_NAME = indCols.INDEX_NAME
-        AND ind.TABLE_OWNER = indCols.TABLE_OWNER
-    LEFT JOIN
-        ALL_CONSTRAINTS cons ON cons.INDEX_NAME = ind.INDEX_NAME
-        AND cons.OWNER = indCols.TABLE_OWNER
-        AND cons.CONSTRAINT_TYPE = 'P'
-    WHERE
-        ind.UNIQUENESS = 'UNIQUE'
-        AND cons.CONSTRAINT_TYPE IS NULL
-        AND ind.TABLE_OWNER = '%s'
-        AND indCols.TABLE_NAME IN ('%s')
-)
-SELECT TABLE_NAME, index_key, COLUMN_NAME, ordinal_position FROM unique_constraints
-UNION ALL
-SELECT TABLE_NAME, index_key, COLUMN_NAME, ordinal_position FROM unique_indexes
-ORDER BY TABLE_NAME, index_key, ordinal_position
-`
-
-func buildUniqueIndexesMapFromOracleRows(rows *sql.Rows, tableStrToNameTupleMap map[string]sqlname.NameTuple) (*utils.StructMap[sqlname.NameTuple, [][]string], error) {
-	indexColumnMaps := make(map[tableIndexKey]map[int]string)
-
-	for rows.Next() {
-		var tableName, indexKey, columnName string
-		var ordinalPosition int
-		err := rows.Scan(&tableName, &indexKey, &columnName, &ordinalPosition)
-		if err != nil {
-			return nil, fmt.Errorf("scanning row for unique index column: %w", err)
-		}
-		key := tableIndexKey{tableCatalogName: tableName, indexKey: indexKey}
-		if _, ok := indexColumnMaps[key]; !ok {
-			indexColumnMaps[key] = make(map[int]string)
-		}
-		indexColumnMaps[key][ordinalPosition] = columnName
-	}
-	if err := rows.Err(); err != nil {
-		return nil, fmt.Errorf("iterating rows for unique indexes: %w", err)
-	}
-
-	result := utils.NewStructMap[sqlname.NameTuple, [][]string]()
-	for key, ordinalToColumn := range indexColumnMaps {
-		tableNameTuple, ok := tableStrToNameTupleMap[key.tableCatalogName]
-		if !ok {
-			return nil, goerrors.Errorf("table %s not found in table list", key.tableCatalogName)
-		}
-
-		ordinals := lo.Keys(ordinalToColumn)
-		sort.Ints(ordinals)
-		columns := make([]string, 0, len(ordinals))
-		for _, ord := range ordinals {
-			columns = append(columns, ordinalToColumn[ord])
-		}
-		if len(columns) == 0 {
-			continue
-		}
-
-		indexes, _ := result.Get(tableNameTuple)
-		indexes = append(indexes, []string(columns))
-		result.Put(tableNameTuple, dedupeUniqueIndexes(indexes))
-	}
-
-	return result, nil
-}
-
-func queryOracleUniqueIndexesMap(db *sql.DB, owner string, tableList []sqlname.NameTuple) (*utils.StructMap[sqlname.NameTuple, [][]string], error) {
-	tableStrToNameTupleMap := make(map[string]sqlname.NameTuple)
-	var queryTableList []string
-	for _, table := range tableList {
-		_, tname := table.ForCatalogQuery()
-		queryTableList = append(queryTableList, tname)
-		tableStrToNameTupleMap[tname] = table
-	}
-
-	query := fmt.Sprintf(oraQueryTmplForUniqIndexes, owner, strings.Join(queryTableList, "','"), owner, strings.Join(queryTableList, "','"))
-	rows, err := db.Query(query)
-	if err != nil {
-		return nil, fmt.Errorf("querying unique indexes: %w", err)
-	}
-	defer rows.Close()
-
-	return buildUniqueIndexesMapFromOracleRows(rows, tableStrToNameTupleMap)
-}
-
-func (ora *Oracle) GetTableToUniqueIndexesMap(tableList []sqlname.NameTuple) (*utils.StructMap[sqlname.NameTuple, [][]string], error) {
-	log.Infof("getting unique indexes for tables: %v", tableList)
-	result, err := queryOracleUniqueIndexesMap(ora.db, ora.source.Schemas[0].Unquoted, tableList)
-	if err != nil {
-		return nil, err
-	}
-	log.Infof("unique indexes for tables: %+v", result)
-	return result, nil
 }
 
 const DROP_TABLE_IF_EXISTS_QUERY = `BEGIN
