+++ yb-voyager/src/srcdb/yugabytedb.go
@@ -1034,87 +1034,6 @@ WHERE parent.relname='%s' AND nmsp_parent.nspname = '%s' `, tname, sname)
 	return partitions
 }
 
-// query retrieves unique indexes/constraints with column order for PG-compatible catalogs.
-const pgQueryTmplForUniqIndexes = `
-WITH unique_constraints AS (
-    SELECT
-        tc.table_schema,
-        tc.table_name,
-        tc.constraint_name AS index_key,
-        kcu.column_name,
-        kcu.ordinal_position
-    FROM
-        information_schema.table_constraints tc
-    JOIN
-        information_schema.key_column_usage kcu
-        ON tc.constraint_name = kcu.constraint_name
-        AND tc.table_schema = kcu.table_schema
-        AND tc.table_name = kcu.table_name
-    WHERE
-        tc.constraint_type = 'UNIQUE'
-        AND tc.table_schema = ANY('{%s}')
-        AND tc.table_name = ANY('{%s}')
-),
-unique_indexes AS (
-    SELECT
-        n.nspname AS table_schema,
-        t.relname AS table_name,
-        i.relname AS index_key,
-        a.attname AS column_name,
-        array_position(ix.indkey, a.attnum) + 1 AS ordinal_position
-    FROM
-        pg_index ix
-    JOIN
-        pg_class i ON i.oid = ix.indexrelid
-    JOIN
-        pg_class t ON t.oid = ix.indrelid
-    JOIN
-        pg_namespace n ON n.oid = t.relnamespace
-    JOIN
-        pg_attribute a ON a.attrelid = t.oid
-        AND a.attnum = ANY(ix.indkey)
-    LEFT JOIN
-        pg_constraint c ON ix.indexrelid = c.conindid AND c.contype = 'p'
-    WHERE
-        ix.indisunique = TRUE
-        AND c.contype IS NULL
-        AND n.nspname = ANY('{%s}')
-        AND t.relname = ANY('{%s}')
-),
-all_unique_indexes AS (
-    SELECT
-        table_schema,
-        table_name,
-        index_key,
-        column_name,
-        MIN(ordinal_position) AS ordinal_position
-    FROM (
-        SELECT table_schema, table_name, index_key, column_name, ordinal_position FROM unique_constraints
-        UNION
-        SELECT table_schema, table_name, index_key, column_name, ordinal_position FROM unique_indexes
-    ) AS combined
-    GROUP BY table_schema, table_name, index_key, column_name
-)
-SELECT
-    table_schema,
-    table_name,
-    index_key,
-    array_agg(column_name ORDER BY ordinal_position) AS columns
-FROM all_unique_indexes
-GROUP BY table_schema, table_name, index_key
-ORDER BY table_schema, table_name, index_key;
-`
-
-func (yb *YugabyteDB) GetTableToUniqueIndexesMap(tableList []sqlname.NameTuple) (*utils.StructMap[sqlname.NameTuple, [][]string], error) {
-	log.Infof("getting unique indexes for tables: %v", tableList)
-	result, err := queryPGUniqueIndexesMap(yb.db, tableList)
-	if err != nil {
-		return nil, err
-	}
-	log.Infof("unique indexes for tables: %v", result)
-	return result, nil
-}
-
 func (yb *YugabyteDB) ClearMigrationState(migrationUUID uuid.UUID, exportDir string) error {
 	log.Infof("ClearMigrationState not implemented yet for YugabyteDB")
 	return nil
