+++ yb-voyager/src/tgtdb/postgres.go
@@ -379,6 +379,175 @@ func (pg *TargetPostgreSQL) GetPrimaryKeyConstraintNames(table sqlname.NameTuple
 	return nil, nil
 }
 
+func (pg *TargetPostgreSQL) GetTableToUniqueIndexesMap(tableList []sqlname.NameTuple) (*utils.StructMap[sqlname.NameTuple, [][]string], error) {
+	log.Infof("getting unique indexes from target for tables: %v", tableList)
+	result, err := queryPGUniqueIndexesMap(pg.Query, tableList)
+	if err != nil {
+		return nil, err
+	}
+	log.Infof("unique indexes from target for tables: %v", result)
+	return result, nil
+}
+
+// pgQueryTmplForUniqIndexes returns, for the given schema/table lists, every unique
+// constraint and unique index (excluding primary keys) with its ordered column list.
+// It is used for both PostgreSQL and YugabyteDB targets since YugabyteDB is
+// PG-compatible at the catalog level.
+const pgQueryTmplForUniqIndexes = `
+WITH unique_constraints AS (
+    SELECT
+        tc.table_schema,
+        tc.table_name,
+        tc.constraint_name AS index_key,
+        kcu.column_name,
+        kcu.ordinal_position
+    FROM
+        information_schema.table_constraints tc
+    JOIN
+        information_schema.key_column_usage kcu
+        ON tc.constraint_name = kcu.constraint_name
+        AND tc.table_schema = kcu.table_schema
+        AND tc.table_name = kcu.table_name
+    WHERE
+        tc.constraint_type = 'UNIQUE'
+        AND tc.table_schema = ANY('{%s}')
+        AND tc.table_name = ANY('{%s}')
+),
+unique_indexes AS (
+    SELECT
+        n.nspname AS table_schema,
+        t.relname AS table_name,
+        i.relname AS index_key,
+        a.attname AS column_name,
+        array_position(ix.indkey, a.attnum) + 1 AS ordinal_position
+    FROM
+        pg_index ix
+    JOIN
+        pg_class i ON i.oid = ix.indexrelid
+    JOIN
+        pg_class t ON t.oid = ix.indrelid
+    JOIN
+        pg_namespace n ON n.oid = t.relnamespace
+    JOIN
+        pg_attribute a ON a.attrelid = t.oid
+        AND a.attnum = ANY(ix.indkey)
+    LEFT JOIN
+        pg_constraint c ON ix.indexrelid = c.conindid AND c.contype = 'p'
+    WHERE
+        ix.indisunique = TRUE
+        AND c.contype IS NULL
+        AND n.nspname = ANY('{%s}')
+        AND t.relname = ANY('{%s}')
+),
+all_unique_indexes AS (
+    SELECT
+        table_schema,
+        table_name,
+        index_key,
+        column_name,
+        MIN(ordinal_position) AS ordinal_position
+    FROM (
+        SELECT table_schema, table_name, index_key, column_name, ordinal_position FROM unique_constraints
+        UNION
+        SELECT table_schema, table_name, index_key, column_name, ordinal_position FROM unique_indexes
+    ) AS combined
+    GROUP BY table_schema, table_name, index_key, column_name
+)
+SELECT
+    table_schema,
+    table_name,
+    index_key,
+    array_agg(column_name ORDER BY ordinal_position) AS columns
+FROM all_unique_indexes
+GROUP BY table_schema, table_name, index_key
+ORDER BY table_schema, table_name, index_key;
+`
+
+// dedupeUniqueIndexes removes duplicate index definitions, keyed by the ordered
+// column signature.
+func dedupeUniqueIndexes(indexes [][]string) [][]string {
+	seen := make(map[string]bool)
+	result := make([][]string, 0, len(indexes))
+	for _, cols := range indexes {
+		key := strings.Join(cols, ",")
+		if seen[key] {
+			continue
+		}
+		seen[key] = true
+		result = append(result, cols)
+	}
+	return result
+}
+
+// queryPGUniqueIndexesMap runs the PG/YB unique-index discovery query for the given
+// tables using the provided query function and returns a map from table to its list
+// of unique indexes (each represented as an ordered column list).
+func queryPGUniqueIndexesMap(queryFn func(query string) (*sql.Rows, error), tableList []sqlname.NameTuple) (*utils.StructMap[sqlname.NameTuple, [][]string], error) {
+	tableStrToNameTupleMap := make(map[string]sqlname.NameTuple)
+	var querySchemaList, queryTableList []string
+	for _, table := range tableList {
+		sname, tname := table.ForCatalogQuery()
+		querySchemaList = append(querySchemaList, sname)
+		queryTableList = append(queryTableList, tname)
+		tableStrToNameTupleMap[table.AsQualifiedCatalogName()] = table
+	}
+	querySchemaList = lo.Uniq(querySchemaList)
+
+	query := fmt.Sprintf(pgQueryTmplForUniqIndexes,
+		strings.Join(querySchemaList, ","), strings.Join(queryTableList, ","),
+		strings.Join(querySchemaList, ","), strings.Join(queryTableList, ","))
+	rows, err := queryFn(query)
+	if err != nil {
+		return nil, fmt.Errorf("querying unique indexes: %w", err)
+	}
+	defer func() {
+		closeErr := rows.Close()
+		if closeErr != nil {
+			log.Errorf("closing rows for unique indexes: %v", closeErr)
+		}
+	}()
+
+	return buildUniqueIndexesMapFromPGRows(rows, tableStrToNameTupleMap)
+}
+
+func buildUniqueIndexesMapFromPGRows(rows *sql.Rows, tableStrToNameTupleMap map[string]sqlname.NameTuple) (*utils.StructMap[sqlname.NameTuple, [][]string], error) {
+	result := utils.NewStructMap[sqlname.NameTuple, [][]string]()
+
+	for rows.Next() {
+		var schemaName, tableName, indexKey string
+		var columnsPgTypeArray pgtype.TextArray
+		err := rows.Scan(&schemaName, &tableName, &indexKey, &columnsPgTypeArray)
+		if err != nil {
+			return nil, fmt.Errorf("scanning row for unique index: %w", err)
+		}
+		columns := utils.ConvertPgTextArrayToStringSlice(columnsPgTypeArray)
+		if len(columns) == 0 {
+			continue
+		}
+
+		tableCatalogName := fmt.Sprintf("%s.%s", schemaName, tableName)
+		tableNameTuple, ok := tableStrToNameTupleMap[tableCatalogName]
+		if !ok {
+			return nil, goerrors.Errorf("table %s not found in table list", tableCatalogName)
+		}
+
+		indexes, _ := result.Get(tableNameTuple)
+		indexes = append(indexes, columns)
+		result.Put(tableNameTuple, indexes)
+	}
+	if err := rows.Err(); err != nil {
+		return nil, fmt.Errorf("iterating rows for unique indexes: %w", err)
+	}
+
+	result.IterKV(func(k sqlname.NameTuple, v [][]string) (bool, error) {
+		v = dedupeUniqueIndexes(v)
+		result.Put(k, v)
+		return true, nil
+	})
+
+	return result, nil
+}
+
 func (pg *TargetPostgreSQL) GetNonEmptyTables(tables []sqlname.NameTuple) []sqlname.NameTuple {
 	result := []sqlname.NameTuple{}
